Split CustomFieldsForm subcomponents
